# roaring: iterate over containers directly